Changed from datasets to polars usage. Merged star download and processing.
