+++ pyproject.toml
@@ -5,12 +5,15 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
-    "datasets>=4.4.1",
+    "huggingface-hub>=1.1.4",
     "librosa>=0.11.0",
     "lxml>=6.0.2",
     "matplotlib>=3.10.7",
+    "mlcroissant>=1.0.22",
     "pandas>=2.3.3",
+    "polars>=1.35.2",
     "pyarrow>=22.0.0",
+    "requests>=2.32.5",
     "tdqm>=0.0.1",
     "wfdb>=4.3.0",
     "wikipedia>=1.4.0",
