+++ src/data/star/download-star.py
@@ -1,8 +1,8 @@
 # Purpose:
-# Retrieves S&P 500 stock tickers from Wikipedia.
-# Downloads stock data from January 1st 2000 to November 1st 2025 for all retrieved tickers
-# Computes daily simple return of stocks
-# Saves return files to parquet
+# Retrieves StarEmbed dataset.
+# Merges training and validation set.
+# Retrieves light wave and timestamp data.
+# Saves datasets.
 
 # Authors:
 # - Code written by Philip Loewen
@@ -15,28 +15,82 @@
 # StarEmbed: Benchmarking Time Series Foundation models on astronomical observations of
 # variable stars. arXiv (Cornell University). https://doi.org/10.48550/arxiv.2510.06200
 
-from datasets import load_dataset
+import polars as pl
 import pandas as pd
+import numpy as np
 import os
+
+
+# Creates a dictionary entry for light colour c and observation i
+def add_observation_to_dict(dict, c, i_data):
+    c_data = i_data[c]
+    target_str = c + "_target"
+    mjd_str = c + "_mjd"
+
+    if c_data is None:
+        dict[target_str] = np.zeros(300)
+        dict[mjd_str] = np.array(range(300))
+    else:
+        dict[target_str] = c_data["target"]
+        dict[mjd_str] = c_data["mjd"]
+
+
+# Builds a dictionary of input variables for data for one split
+def build_dict_for_split(data):
+    dict = {}
+    n = data.shape[0]
+    bands = data["bands_data"]
+    for i in range(n):
+        d = dict[i] = {}
+        i_data = bands[i]
+        for c in ("g", "r"):
+            add_observation_to_dict(d, c, i_data)
+
+    return dict
+
+
+# Login using e.g. `huggingface-cli login` to access this dataset
+splits = {
+    "train": "data/train-*.parquet",
+    "validation": "data/validation-00000-of-00001.parquet",
+    "test": "data/test-00000-of-00001.parquet",
+}
 
 # Download huggingface dataset
 print("Downloading data...")
-ds = load_dataset("123anonymous123/StarEmbed")
-print("Done downloading data.")
+train = pl.read_parquet("hf://datasets/123anonymous123/StarEmbed/" + splits["train"])
+test = pl.read_parquet("hf://datasets/123anonymous123/StarEmbed/" + splits["test"])
+val = pl.read_parquet("hf://datasets/123anonymous123/StarEmbed/" + splits["validation"])
+
+print("Making dataframes...")
 
 # If raw data folder does not exist, create it
-output_path = "data/raw/star"
+output_path = "data/processed/star"
 if not os.path.exists(output_path):
     os.makedirs(output_path)
 
-print("Saving data...")
-train_data = pd.DataFrame(ds["train"])
-val_data = pd.DataFrame(ds["validation"])
-merged_data = pd.concat([train_data, val_data])
+colnames = train.columns
 
-merged_data.to_parquet(os.path.join(output_path, "train-star-raw.parquet"))
+# Merge training and validation sets
+merged_train = train.vstack(val)
+merged_data = pd.DataFrame(build_dict_for_split(merged_train)).transpose()
+test_data = pd.DataFrame(build_dict_for_split(test)).transpose()
+
+# Get star classes
+y_merge = pd.DataFrame(merged_train["class_str"], columns=["Class"])
+y_test = pd.DataFrame(test["class_str"], columns=["Class"])
+
+
+print("Saving data...")
+
+# Save training data
+merged_data.to_parquet(os.path.join(output_path, "star_X_train.parquet"))
+y_merge.to_parquet(os.path.join(output_path, "star_y_train.parquet"))
 print("Saved training data.")
-ds["test"].to_parquet(os.path.join(output_path, "test-star-raw.parquet"))
+
+# Save testing data
+test_data.to_parquet(os.path.join(output_path, "star_X_test.parquet"))
+y_test.to_parquet(os.path.join(output_path, "star_y_test.parquet"))
 print("Saved testing data.")
 
 print("Done saving data.")
